# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
      "ArmCreateIn": {
        "properties": {
          "class_level_id": {
            "format": "uuid",
            "title": "Class Level Id",
            "type": "string"
          },
          "name": {
            "maxLength": 30,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "class_level_id",
          "name"
        ],
        "title": "ArmCreateIn",
        "type": "object"
      },
...
      "AssessmentMode": {
        "enum": [
          "scored",
          "developmental",
          "mixed"
        ],
        "title": "AssessmentMode",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HouseCreateIn": {
        "properties": {
          "name": {
            "maxLength": 50,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "HouseCreateIn",
        "type": "object"
      },
      "HouseOut": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "HouseOut",
        "type": "object"
      },
...
      "LevelOut": {
        "properties": {
          "arms": {
            "items": {
              "$ref": "#/components/schemas/ArmOut"
            },
            "title": "Arms",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "next_level_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Level Id"
          },
          "sort": {
            "title": "Sort",
            "type": "integer"
          },
          "subject_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Subject Ids",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "sort",
          "next_level_id",
          "arms",
          "subject_ids"
        ],
        "title": "LevelOut",
        "type": "object"
      },
      "LevelUpdateIn": {
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "LevelUpdateIn",
        "type": "object"
      },
...
      "SectionKind": {
        "enum": [
          "creche",
          "nursery",
          "primary",
          "junior_secondary",
          "senior_secondary",
          "other"
        ],
        "title": "SectionKind",
        "type": "string"
      },
      "SectionOut": {
        "properties": {
          "assessment_mode": {
            "$ref": "#/components/schemas/AssessmentMode"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/SectionKind"
          },
          "levels": {
            "items": {
              "$ref": "#/components/schemas/LevelOut"
            },
            "title": "Levels",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "student_login_enabled": {
            "title": "Student Login Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "display_name",
          "kind",
          "assessment_mode",
          "student_login_enabled",
          "levels"
        ],
        "title": "SectionOut",
        "type": "object"
      },
      "SectionUpdateIn": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "student_login_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Login Enabled"
          }
        },
        "title": "SectionUpdateIn",
        "type": "object"
      },
      "SessionCreateIn": {
        "properties": {
          "name": {
            "examples": [
              "2027/2028"
            ],
            "pattern": "^\\d{4}/\\d{4}$",
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "SessionCreateIn",
        "type": "object"
      },
...
      "SessionOut": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "terms": {
            "items": {
              "$ref": "#/components/schemas/TermOut"
            },
            "title": "Terms",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "terms"
        ],
        "title": "SessionOut",
        "type": "object"
      },
      "SetupOverviewOut": {
        "description": "Everything the setup screens need in one request.",
        "properties": {
          "active_session_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Session Id"
          },
          "houses": {
            "items": {
              "$ref": "#/components/schemas/HouseOut"
            },
            "title": "Houses",
            "type": "array"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/SectionOut"
            },
            "title": "Sections",
            "type": "array"
          },
          "sessions": {
            "items": {
              "$ref": "#/components/schemas/SessionOut"
            },
            "title": "Sessions",
            "type": "array"
          },
          "subjects": {
            "items": {
              "$ref": "#/components/schemas/SubjectOut"
            },
            "title": "Subjects",
            "type": "array"
          }
        },
        "required": [
          "sections",
          "sessions",
          "active_session_id",
          "houses",
          "subjects"
        ],
        "title": "SetupOverviewOut",
        "type": "object"
      },
...
      "SubjectIn": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "level_ids": {
            "default": [],
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Level Ids",
            "type": "array"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "SubjectIn",
        "type": "object"
      },
      "SubjectOut": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "code"
        ],
        "title": "SubjectOut",
        "type": "object"
      },
      "TermOut": {
        "properties": {
          "ends_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends On"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_current": {
            "title": "Is Current",
            "type": "boolean"
          },
          "next_term_begins": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Term Begins"
          },
          "number": {
            "title": "Number",
            "type": "integer"
          },
          "starts_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts On"
          }
        },
        "required": [
          "id",
          "number",
          "starts_on",
          "ends_on",
          "next_term_begins",
          "is_current"
        ],
        "title": "TermOut",
        "type": "object"
      },
      "TermUpdateIn": {
        "properties": {
          "ends_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends On"
          },
          "next_term_begins": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Term Begins"
          },
          "starts_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts On"
          }
        },
        "title": "TermUpdateIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "DigitalLearning360 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/setup/arms": {
      "post": {
        "operationId": "setup_create_arm",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArmCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupOverviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Arm",
        "tags": [
          "setup"
        ]
      }
    },
    "/api/setup/arms/{arm_id}": {
      "delete": {
        "operationId": "setup_delete_arm",
        "parameters": [
          {
            "in": "path",
            "name": "arm_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Arm Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupOverviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Arm",
        "tags": [
          "setup"
        ]
      }
    },
    "/api/setup/houses": {
      "post": {
        "operationId": "setup_create_house",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HouseCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupOverviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create House",
        "tags": [
          "setup"
        ]
      }
    },
    "/api/setup/houses/{house_id}": {
      "delete": {
        "operationId": "setup_delete_house",
        "parameters": [
          {
            "in": "path",
            "name": "house_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "House Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupOverviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete House",
        "tags": [
          "setup"
        ]
      }
    },
    "/api/setup/levels/{level_id}": {
      "patch": {
        "operationId": "setup_rename_level",
        "parameters": [
          {
            "in": "path",
            "name": "level_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Level Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LevelUpdateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupOverviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename Level",
        "tags": [
          "setup"
        ]
      }
    },
...
    "/api/setup/sections/{section_id}": {
      "patch": {
        "operationId": "setup_update_section",
        "parameters": [
          {
            "in": "path",
            "name": "section_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Section Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SectionUpdateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupOverviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Section",
        "tags": [
          "setup"
        ]
      }
    },
    "/api/setup/sessions": {
      "post": {
        "operationId": "setup_create_session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupOverviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Session",
        "tags": [
          "setup"
        ]
      }
    },
    "/api/setup/subjects": {
      "post": {
        "operationId": "setup_create_subject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubjectIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupOverviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Subject",
        "tags": [
          "setup"
        ]
      }
    },
    "/api/setup/subjects/{subject_id}": {
      "delete": {
        "operationId": "setup_delete_subject",
        "parameters": [
          {
            "in": "path",
            "name": "subject_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Subject Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupOverviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Subject",
        "tags": [
          "setup"
        ]
      },
      "put": {
        "operationId": "setup_update_subject",
        "parameters": [
          {
            "in": "path",
            "name": "subject_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Subject Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubjectIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupOverviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Subject",
        "tags": [
          "setup"
        ]
      }
    },
    "/api/setup/terms/{term_id}": {
      "patch": {
        "operationId": "setup_update_term",
        "parameters": [
          {
            "in": "path",
            "name": "term_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Term Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TermUpdateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupOverviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Term",
        "tags": [
          "setup"
        ]
      }
    },
    "/api/setup/terms/{term_id}/make-current": {
      "post": {
        "operationId": "setup_make_term_current",
        "parameters": [
          {
            "in": "path",
            "name": "term_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Term Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupOverviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Make Term Current",
        "tags": [
          "setup"
        ]
      }
    },
...
  }
}